+++ pepper/rc/changelog.md
@@ -13,6 +13,8 @@
 - removed `alias` command since it's now possible to replicate its behavior by creating a new command that calls the aliased command and use the `@arg()` expansion
 - removed `find-file` and `find-command` commands as they're now implementable using other builtin commands (see `default_commands.pepper` for an example)
 - removed the old 255 cursor count limit
+- exiting search mode will fully restore the previous cursor state
+- it's now possible to use use the search mode to expand selections
 
 ## 0.24.0
 - handle buffer paths beginning with `./` (on `Buffer::set_path` and `Buffer::find_with_path`)